Mon, Jul 10, 2023, 8:22 PM +03:00

+++ filter-more/helpers.c
@@ -15,7 +15,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
             int num_red = image[i][j].rgbtRed; // -> 0x00 or 0xff
             int num_green = image[i][j].rgbtGreen;
             int num_blue = image[i][j].rgbtBlue;
-            int num_avg = (int) ((num_red + num_green + num_blue) / 3);
+            int num_avg = (int) double((num_red + num_green + num_blue) / 3);
             // printf("%d", num_avg);
             image[i][j].rgbtRed = num_avg;
             image[i][j].rgbtGreen = num_avg;
